scripts/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.3
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
